back_drf/lily_api/utils/serializers.py
from rest_framework import serializers
from orders.models import EmailSubscription


class EmailSerializer(serializers.ModelSerializer):
    class Meta:
        model = EmailSubscription
        fields = ["email"]
