back_drf/requirements.txt
Django
djangorestframework
django-cors-headers === 4.3.1
djangorestframework-simplejwt === 5.3.1
redis
pillow==10.3.0
mysqlclient==2.2.4
stripe